chore(repo): синхронизировать тесты и инфраструктурный baseline
Обновить runtime и bridge-срез WebDriver по текущему состоянию репозитория.
Расширить JavaScript source-generation scaffold и связанные контракты.
Добавить служебные workflow и скрипты для упаковки, диагностики и воспроизводимых прогонов.
Актуализировать guard-проверки и репозиторные артефакты вокруг legacy reference.

+++ Framework/Atom.Web.Services.Markets/Polymarket/PolymarketStrategy.cs
@@ -60,7 +60,7 @@ public sealed class PolymarketTradeSignal
     public long TimestampTicks { get; init; } = Environment.TickCount64;
 
     // Явная реализация IMarketTradeSignal.Action (конверсия енумов)
-    TradeAction IMarketTradeSignal.Action => (TradeAction)(byte)Action;
+    TradeAction IMarketTradeSignal.Action => (TradeAction)Action;
 }
 
 /// <summary>
@@ -154,8 +154,10 @@ public abstract class PolymarketStrategyBase : IPolymarketStrategy
     /// </summary>
     protected PolymarketTradeSignal HoldSignal(string assetId, string reason) => new()
     {
-        AssetId = assetId, Action = PolymarketTradeAction.Hold,
-        Quantity = 0, Reason = reason
+        AssetId = assetId,
+        Action = PolymarketTradeAction.Hold,
+        Quantity = 0,
+        Reason = reason
     };
 
     /// <inheritdoc />
